pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.14",
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "events>=0.5",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "logger>=1.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "slack-sdk>=3.35.0",
    "sqlalchemy>=2.0.40",
    "flask-login>=0.6.3",
    "twilio>=9.5.1",
]
